Refining of roles between AwsClientBuilderUtils and CommonConfig.
AwsClientBuilderUtils felt like it was making configuration decision (like the AWS credentials) when it's intended roles looked like it should be applying configuration to AWS clients (withAWSCredentials).

Removing knowledge of isDev in KinesisReceiverConfig feels like this is the right track.
The hard coded eu-region-1 which is then override by the extending CommonConfig looks suspisious as well.
