+++ common-lib/src/main/scala/com/gu/mediaservice/lib/aws/AwsClientBuilderUtils.scala
@@ -1,27 +1,15 @@
 package com.gu.mediaservice.lib.aws
 
-import com.amazonaws.auth.profile.ProfileCredentialsProvider
-import com.amazonaws.auth.{AWSCredentialsProvider, AWSCredentialsProviderChain, InstanceProfileCredentialsProvider}
+import com.amazonaws.auth.AWSCredentialsProvider
 import com.amazonaws.client.builder.AwsClientBuilder
 import com.amazonaws.client.builder.AwsClientBuilder.EndpointConfiguration
 import com.gu.mediaservice.lib.logging.GridLogging
+
 trait AwsClientBuilderUtils extends GridLogging {
-  def awsLocalEndpoint: Option[String]
-  def isDev: Boolean
 
-  def awsRegion: String = "eu-west-1"
-
-  // TODO Config extending this class to advertise withAWSCredentials is proving quite immuteable;
-  // awsCredentials and awsEndpointConfiguration likewise feel like config functions
-  def awsCredentials: AWSCredentialsProvider = new AWSCredentialsProviderChain(
-    new ProfileCredentialsProvider("media-service"),
-    InstanceProfileCredentialsProvider.getInstance()
-  )
-
-  final def awsEndpointConfiguration: Option[EndpointConfiguration] = awsLocalEndpoint match {
-    case Some(endpoint) if isDev => Some(new EndpointConfiguration(endpoint, awsRegion))
-    case _ => None
-  }
+  def awsEndpointConfiguration: Option[EndpointConfiguration]
+  def awsCredentials: AWSCredentialsProvider
+  def awsRegion: String
 
   final def withAWSCredentials[T, S <: AwsClientBuilder[S, T]](builder: AwsClientBuilder[S, T], localstackAware: Boolean = true): S = {
     awsEndpointConfiguration match {
@@ -32,4 +20,5 @@ trait AwsClientBuilderUtils extends GridLogging {
       case _ => builder.withCredentials(awsCredentials).withRegion(awsRegion)
     }
   }
+
 }
