+++ common-lib/src/main/scala/com/gu/mediaservice/lib/aws/Kinesis.scala
@@ -1,24 +1,23 @@
 package com.gu.mediaservice.lib.aws
 
-import java.nio.ByteBuffer
-import java.util.UUID
-
+import com.amazonaws.auth.AWSCredentialsProvider
+import com.amazonaws.client.builder.AwsClientBuilder.EndpointConfiguration
 import com.amazonaws.services.kinesis.model.PutRecordRequest
 import com.amazonaws.services.kinesis.{AmazonKinesis, AmazonKinesisClientBuilder}
-import com.gu.mediaservice.lib.config.CommonConfig
 import com.gu.mediaservice.lib.json.JsonByteArrayUtil
+import com.gu.mediaservice.lib.logging.GridLogging
 import com.gu.mediaservice.model.usage.UsageNotice
 import net.logstash.logback.marker.{LogstashMarker, Markers}
-import play.api.libs.json.{JodaWrites, Json, Writes}
-import com.amazonaws.auth.AWSCredentialsProvider
-import com.gu.mediaservice.lib.logging.GridLogging
 import org.joda.time.DateTime
+import play.api.libs.json.{JodaWrites, Json, Writes}
+
+import java.nio.ByteBuffer
+import java.util.UUID
 
 case class KinesisSenderConfig(
   override val awsRegion: String,
   override val awsCredentials: AWSCredentialsProvider,
-  override val awsLocalEndpoint: Option[String],
-  override val isDev: Boolean,
+  override val awsEndpointConfiguration: Option[EndpointConfiguration],
   streamName: String
 ) extends AwsClientBuilderUtils
